move CoordinateProposalType to /domain
