+++ test/domain/routing.test.js
@@ -1,4 +1,4 @@
-import { RoutingStatusCodes } from '../../src/domain/routing';
+import { CoordinateProposalType, RoutingStatusCodes } from '../../src/domain/routing';
 
 describe('RoutingStatusCodes', () => {
 	it('provides an enum of all valid status codes', () => {
@@ -10,5 +10,18 @@ describe('RoutingStatusCodes', () => {
 		expect(RoutingStatusCodes.Start_Destination_Missing).toBe(900);
 		expect(RoutingStatusCodes.Destination_Missing).toBe(901);
 		expect(RoutingStatusCodes.Start_Missing).toBe(902);
+	});
+});
+
+describe('CoordinateProposalType', () => {
+	it('provides an enum of all possible proposal types', () => {
+		expect(Object.keys(CoordinateProposalType).length).toBe(6);
+		expect(Object.isFrozen(CoordinateProposalType)).toBeTrue();
+		expect(CoordinateProposalType.START_OR_DESTINATION).toBe(0);
+		expect(CoordinateProposalType.START).toBe(1);
+		expect(CoordinateProposalType.DESTINATION).toBe(2);
+		expect(CoordinateProposalType.INTERMEDIATE).toBe(3);
+		expect(CoordinateProposalType.EXISTING_START_OR_DESTINATION).toBe(4);
+		expect(CoordinateProposalType.EXISTING_INTERMEDIATE).toBe(5);
 	});
 });
