+++ test/store/routing/routing.reducer.test.js
@@ -1,6 +1,5 @@
-import { RoutingStatusCodes } from '../../../src/domain/routing';
+import { CoordinateProposalType, RoutingStatusCodes } from '../../../src/domain/routing';
 import {
-	CoordinateProposalType,
 	activate,
 	deactivate,
 	removeWaypoint,
